Policies and reposnses are added

+++ app/Http/Controllers/ContactsController.php
@@ -4,27 +4,41 @@ namespace App\Http\Controllers;
 
 use App\Contact;
 use App\Http\Requests\ContactRequest;
-use Illuminate\Http\Request;
+use App\Http\Resources\Contact as ContactResource;
+use Symfony\Component\HttpFoundation\Response;
 
 class ContactsController extends Controller
 {
+    public function index()
+    {
+        $this->authorize('viewAny', Contact::class);
+        return ContactResource::collection(request()->user()->contacts);
+    }
+
     public function show(Contact $contact)
     {
-        return $contact;
+        $this->authorize('view', $contact);
+        return new ContactResource($contact);
     }
 
     public function store(ContactRequest $request)
     {
-        Contact::create($request->all());
+        $this->authorize('create', Contact::class);
+        $contact = request()->user()->contacts()->create($request->except('api_token'));
+        return (new ContactResource($contact))->response()->setStatusCode(Response::HTTP_CREATED);
     }
 
     public function update(Contact $contact, ContactRequest $request)
     {
-        $contact->update($request->all());
+        $this->authorize('update', $contact);
+        $contact->update($request->except('api_token'));
+        return (new ContactResource($contact))->response()->setStatusCode(Response::HTTP_OK);
     }
 
     public function destroy(Contact $contact)
     {
+        $this->authorize('delete', $contact);
         $contact->delete();
+        return response([], Response::HTTP_NO_CONTENT);
     }
 }
